process-expertise-function/requirements.txt
functions-framework==3.*
youtube-transcript-api
google-cloud-storage
